oracle connection

+++ dotnet/NetCoreLinqToSqlInjection/NetCoreLinqToSqlInjection/Models/Doctor.cs
@@ -6,6 +6,6 @@
         public string DoctorCod { get; set; }
         public string Apellido { get; set; }
         public string Especialidad { get; set; }
-        public int Salario { get; set; }
+        public Decimal Salario { get; set; }
     }
 }
